pam/userselection: Do not draw the model when not enabled
When the model is not enabled we may still have it focused, since it's
the default initial model that we have in the CLI mode.

This may cause to a visual issue (not a security one, since the input
is disabled for this):
 - login with the UI using a preset user
 - while the client is connecting to the server we may show a "frozen"
   user selection view with the pre-selected  user written in.

This is not really something we need to show, so do not draw anything
until we've something to show.