pam/integration-tests/ssh: Ignore SSH tests when running on unsupported OS
SSH output changes from old jammy (where CI is) to noble and greater
versions (as per the OpenSSH server changes we carry on), so to be able
to run the tests in a reliable way we need to be on such context.

While we target noble, we didn't force our CI to be updated, so for now
let's just enable the tests in older ubuntu versions where CI resides
skipping it otherwise.

Added also a further check so that when CI changes we get an error about