pam/integration-tests/vhs-helpers: Add support for Wait+{Suffix,Prompt,Nth}
Simplify the regexes we're using in our tape files by allowing to wait
for a prompt value or for a repeated value in the native model.

This allows to avoid mistakes by easily updating the regexes without
updating the tape files.

Wait+Suffix is also supported as base for Wait+Prompt