pam/integration-tests/gdm: Try doing broker selection also on received brokers
We have a race on broker selection because we may choose a broker when
we don't have one yet, as this depends on the order of the events that
may not always be the same.

To prevent this:
 1. If received the brokers and we're in broker selection stage
    -> Do the broker selection right away, marking it as done
 2. If we just switched to broker selection stage
    a. If the brokers have been received, try to switch to the selected
       one or error if not found.
    b. If brokers have not yet been received, we just ignore the request
       and perform it once the brokers have been provided.

Partially go back to 48db8521 logic.

See: https://github.com/ubuntu/authd/actions/runs/12144656077/job/33864519830