pam/integration-tests: Increase the typing speed
We don't care much about being fast in typing, while we can save some
time in all the tests. And ideally trigger more races 😈!

The VHS default typing speed was of 50ms, that makes the terminal being
rendered nicely if you want to generate gifs with VHS, but that's not
our goal here.

As per the previous fix, all the tests behave as expected while in CI
we'll still test the slow-down cases when using asan or race builds