pam/authmodeselection: Do not focus until authentication is completed
If being fast enough (as when using 1ms typing speed in VHS) we may end
up triggering a race in our authentication model when switching from the
challenge phase to the authentication mode selection, that eventually
caused that no new authentication was re-started after switching mode.

In fact what it could happen is:
 - User selects a Wait authentication mode
 - The user quickly goes back hitting Escape

At this point, as per the implicit authmode selection stage request:
 - We reset the authentication model
   => thus cancelling the authentication in progress
 - We focus the auth mode selection model

In theory, all this should already have happened in order, since we use
a tea.Sequence of events, and then the authentication model should have
been focused only after that the cancellation took place, but this was
not the case because the authModeSelection Focus() method didn't really
lead to any event, but instead was applying the change immediately,
causing the authentication selection to start when we didn't finish
the previous authentication phase.

To prevent this, use an event to notify that the auth mode selection
should be focused, so that it's served in order only *after* the
authentication cancellation is completed.

See this log for details of the failure (from test
 TestCLIChangeAuthTok/Change_passwd_after_MFA_auth):

  adapter.ChangeStage{Stage:3}
  *adapter.authModeSelectionModel: Blur
  *adapter.authenticationModel: Focus
  adapter.formModel: Focus
  *adapter.authenticationModel: adapter.startAuthentication{}:
    current model {Plug your fido device and press with your thumb [] 0 true},
    focused true
  adapter.isAuthenticatedRequested{
    item:(*authd.IARequest_AuthenticationData_Wait)(0xc00041c8e0)}
  adapter.isAuthenticatedRequestedSend{
    isAuthenticatedRequested:adapter.isAuthenticatedRequested{
      item:(*authd.IARequest_AuthenticationData_Wait)(0xc00041c8e0)},
      ctx:(*context.cancelCtx)(0xc000180190)}
  Authentication request for session
    "2221040704-3d60ebd9-4ee2-408e-b3fe-92b622bcc5fb":
      &authd.IARequest_AuthenticationData_Wait{Wait:"true"}
  Key: "esc" in stage "challenge"
  *adapter.authenticationModel: Blur
  adapter.formModel: Blur
  *adapter.authenticationModel: Reset
  *adapter.authModeSelectionModel: Focus
  Key: "enter" in stage "authModeSelection"
  adapter.authModeSelected{id:"fidodevice1"}
  adapter.AuthModeSelected{ID:"fidodevice1"}
  adapter.UILayoutReceived{layout:(*authd.UILayout)(0xc0000242a0)}
  adapter.ChangeStage{Stage:3}
  *adapter.authModeSelectionModel: Blur
  *adapter.authenticationModel: Focus
  adapter.formModel: Focus
  *adapter.authenticationModel: adapter.startAuthentication{}: current model {
    Plug your fido device and press with your thumb [] 0 true}, focused true
  adapter.isAuthenticatedRequested{
    item:(*authd.IARequest_AuthenticationData_Wait)(0xc00041cf50)}
  Authentication completed for session
    "2221040704-3d60ebd9-4ee2-408e-b3fe-92b622bcc5fb":
      adapter.isAuthenticatedResultReceived{
        access:"cancelled", challenge:(*string)(nil), msg:"}
  adapter.isAuthenticatedResultReceived{
    access:"cancelled", challenge:(*string)(nil), msg:"}

The module hanged here, without being able to proceed to the
authentication.