Remove incorrect os.Exit(m.Run()) calls
This is not needed since Go 1.15, where the test runtime registers the
test result itself.
Also, os.Exit() would skip any defer() being called before, which might
pollute the system running the tests.

Note that we only kept 2 occurences when subprocessing as a mock, as we
shouldn't expect to have defer() beforehand in TestMain.