pam: Block until the connection with the daemon is ready
An authd instance running is a strong prerequisite for for the authd PAM
module too, so instead of failing at the first grpc method invocation
with a system error, let's just try to connect first and in case of
failure, we just fail (as expected) with pam.ErrAuthinfoUnavail so that
the service files can decide what to do with it (normally: ignore it).

Add tests.