daemon: Use GRPC HealthCheck instead of manual connection check
GRPC provides an healthcheck service and clients to ensure that specific
services are up and running when required.

Using this instead of connection wait is better because it also allows
us to ensure that the socket we're connecting to:
 1. it's available
 2. it's responding to GRPC queries
 3. it's implementing the required APIs

authd initializes all the services together and starts serving only
after that has setup everything, so we don't need to change the health
state at runtime, but still this allows us to ensure that grpc machinery
is available.