Use lowercase usernames
All the brokers we currently support (msentraid and google) use
case-insensitive usernames. Currently, logging in with a username which
differs in upper- or lowercase from the username in the broker results
in an authentication failure. We want to fix that.

The easiest way to do that is to convert all usernames to lowercase
before storing them in our database or passing them to the broker, and
also convert the Name argument of GetPassdByName to lowercase before
querying the database.

If we (or anyone else) ever want to add a broker for a provider which
does *not* use case-insensitive usernames, we will have to revisit this
and decide whether we want to support that.