Migrate database to lowercase usernames
To be able to treat usernames as case-insensitive, we store them in
lowercase in our database and query them in lowercase. To make that work
with existing databases, we need to convert the usernames of all
existing user records in the database to lowercase.

Note that we don't modify the home directory. If an uppercase username
is part of the home directory, it will still be in uppercase.