Fix login for users which already logged in with previous authd version
We only store the UGID of users since
edd2399ca70972eddad58e355b6897ca4f3c3f1c, which has not been released
yet. Therefore, we also need to search groups by name to avoid that we
try to generate a new GID for them (which involves registering a
temporary name, which fails because the name already exists).