Use "authd group" as the term for the opposite of a local group
The difference between a local and a non-local group is that the local
group is stored in /etc/group and the non-local group is stored in the
authd database, so it makes more sense to call it an "authd group"
instead of a "cloud group". We already use that term in the function
body of UpdateUser.