debian: Ensure that /etc/authd has mode 700
In versions 0.3.7 and earlier, /etc/authd was created mode 0777 - umask
(so 0755 by default). Since 1d8254abdf9a9ed70e189c594d374203bf6f06c1, we
tell systemd that it should create the directory with mode 0700, which
results in the following warning in the journal:

   authd.service: ConfigurationDirectory 'authd' already exists but the mode
   is different. (File system: 755 ConfigurationDirectoryMode: 700)

To avoid that, this commit ensures that the mode is changed to 0700 when
we upgrade from version 0.3.7 or earlier.