debian/control: Add cargo-vendor-filterer in devel version
Ubuntu devel already contains rust-cargo-vendor-filterer, we can't yet
build-depend on it in debian/control for everybody (as it's not in noble
yet), but we can still workaround this.

A change to dh-rust (just after the last run of the
build-deb run on the PR), broke the usage of cargo-vendor-filterer
when installed from cargo itself.

This is ugly, but avoids us to branch for now.