ci/auto-updates: Fail if dh-cargo-vendored-sources fails for unknown reasons
We were handling the failure in case XS-Vendored-Sources-Rust changed,
but in case some other failure is happening (like right now in plucky),
then we would have just ignored it.

So instead of relying only on the grep output, check if the dh-cargo
source vendoring tool failed for some other reason and return about.