Don't delete empty groups from database
When a user logs in, we fetch the groups from the provider and update
our database accordingly. When this causes a user to be removed from an
authd group and that group doesn't have any other users in it, it's
currently removed from our database. That's an issue for two reasons:

1. The next time a user who is a member of that group logs in, a new
   random GID is generated for that group, which means that any existing
   files owned by the group won't be accessible to members of the group
   anymore.

2. Whenever a another group is added, the random GID generated for that
   group can by chance be the same as the GID of the deleted group,
   allowing members access to existing files owned by the deleted group.

With this commit, we don't delete empty groups but just keep them in the
database, so that its GID is still reserved and reused the next time a
user who is a member of that group logs in.