debian/rules: Do not check vendored sources during configure phase
If cargo-vendor-filter is used to generate the debian/control vendored
sources it won't include the windows and other filtered crates.

However the version that is in noble is not ready for this since it does
not support filtering and then we'd just fail at configure phase,
expecting the filtered crates to be there.

As per this, in the cargo filterer missing case, just do manually what
dh-cargo does without the vendoring checks