Don't delete user from database if updating local groups fails
It was broken in multiple ways:

* It deletes the user even if it already existed in the database before
  this update.

* It's not protected by a mutex, so a concurrent update could have
  successfully added the user, which is then incorrectly deleted here.

In addition to that, now that we don't automatically delete empty
groups, it still leaves the groups (including the user private group)
which were added.

Fixing all that would require effort which I'm not sure would be
justified. I don't see any issues with users/groups from an
authenticated but erroneous login attempt still being in the database,
except that it might be a bit surprising to users.

Therefore, with this commit we simply leave the updated user and groups
as they are.