highlight and differentiate Ubuntu systems more clearly
Originally, only the installation command for Desktop was shown, with a
parenthetical note about not needing certain packages if using Server.
This is now replaced with a tabbed interface, which allows the user to
select the appropriate option and copy the appropriate command for their
system.

In addition, I have tried to make it more clear/obvious where possible
that Desktop and Server are both supported.