testutils/daemon: Add option to run authd with a PID file
The pid file is used to both know what is the pid of the daemon in use
but it's also special so that when it gets removed, the daemon itself is
stopped.

We use this behavior, because if the PID would be killed directly from
an external source, then we'd fail on checking the reasons for being
stopped (e.g. the context wouldn't be cancelled). So we still keep the
control on stopping the daemon itself on us.