WIP: Add behave tests
Add end-to-end UI tests that are executed in a VM and leverage the
accessibility bus to simulate user interaction and check the state of
the system and apps.

The tests are defined in the natural-language Gherkin syntax and
executed via Behave, the Python-implementation of Cucumber.

The VM is automatically provisioned via multipass, and multipass is also
used to start and stop the VM and to execute commands in it.

The accessibility bus is shared via VSOCK, allowing high-performant
communication directly from the host.

Steps in the tests can define snapshots, which are automatically
restored when another test uses the same step, allowing to reuse a
well-defined machine state in separate tests, thereby significantly
reducing the total test runtime without requiring dependencies between
tests (i.e. each test can be executed on its own, without requiring
executing any "previous" test to bring the system into the required
state).