pam/adapter/gdmmodel-test: Avoid racing on "gdm-repeated-password" test
We may end up not queuing the events, leading to a failure since the
changes in this branch (racy bubbletea, you know!).

Mostly this is because now we have a timeout that triggers the
GetAuthenticationModesRequested{} event, that implies that the
authentication modes gets reloaded ant the first-one gets auto-selected,
so in the GDM tests we need to ensure that when "auth.Next" is returned
then we must first wait for the new challenge stage, and finally to
switch back to the auth-mode selection.

See: https://github.com/ubuntu/authd/actions/runs/13550269562/job/37879512072