pam/authentication: Add a dummy focusTracker module when using GDM or PAM
When using the gdm or native modules we still use the authentication
model as our data model to track the focused widget status, however
since we don't have any current model set in such case, we end up never
marking the "Authentication" model as ever tracked.

This has implications in the model stage change, since we do use the
focus state to figure out what is the current stage.

In these case add then a very simple focus tracker stage so that the
authentication state can still marked as focused when it's the case.

This should fix some "focus jumps" we may see in some cases (especially
in GDM)