pam/adapter/model: Reset the challenge stage before updating other states
As per previous commit we are now aware all the times when we're in the
authentication phase, so we can safely just reset the authentication
model only after we've a stage change that goes from it to any other
stage.

Note that resetting the authentication also implies waiting for the
previous authentication request being cancelled, reason why it's
important to do every time we switch back from the focus stage and that
we do it before communicating the stage change to other models such as
gdm and cli.