pam/gdmmodel: Do not perform auth cancellation
As per the fact we're now tracking the focus state also when the gdm
model is used (via the focusTracker model) it's not up anymore to us to
perform any cancellation in the GDM model either, since the generic
authentication model handles it for us already, and in the proper order.

As per this, the AuthResult cancelled event is not sent anymore to gdm
unless if that happens because a wait operation got cancelled, as it
happens in the other models too, given that a stage change now
implicitly already cancels the ongoing authentication request.

Thus, update tests accordingly