pam/go-exec/module: Simplify the connection handling in thread
We can safely call g_connection_close multiple times on a valid
connection, while we already have code handling the connection disposal
in action_module_data_cleanup().

At the same time here there's no risk that we're accessing to a disposed
connection in the thread, because at this point the wait-child thread is
running while the action-thread is waiting for it, before disposing the
connection.