pam/go-exec/module: Simplify the connection handling in thread
We can safely call g_connection_close multiple times on a valid
connection, while we already have code handling the connection disposal
in action_module_data_cleanup().

As per this, there's no need to do the connection cleanup when the
wait-child thread is done, nor there's risk that we're accessing to a
disposed connection pointer, because at the point the wait-child thread
is running, the action-thread is waiting for it. Just before disposing
the connection.