pam/go-exec/module: More consistent behavior for fatal signals
Sadly some signals such as SIGABRT or SIGSEGV are handled by go and in
the wrong way because it never redirects them as expected, so in such
cases we just fail with a normal exit error instead of because of a
signal.

Reported this upstream and adding comments about.

See: https://github.com/golang/go/issues/72084