pam/gdmmodel-test: Fix tests where we need to send commands after authentication
In some tests we need to send events or commands juts after the the
authentication data is sent, however we were doing this without a clear
order, because the gdmTestSendAuthDataWhenReady{} commands may have been
completed when the next command was processed.

So use instead the nested commands via gdmTestSendAuthDataWhenReadyFull,
in this way we're sure that the next commands will be processed only
after that we're done with the authentication.

This should fix various tests races