Remove database on schema creation failure
I noticed that if the schema creation fails (e.g. because the schema is
invalid), the database file is still created and authd uses it the next
time it's started. That results in authd starting up successfully but in
failure of each request which uses the broken database.

Let's avoid that by ensuring that the database file is removed when the
schema creation fails.