pam/authmodeselection: Auto-select the first auth mode only once focused
On authModesReceived{} message we were aggressively selecting the first
authentication mode, despite the authentication stage was not being
focused, and this could have lead to troubles because a previously
queued stage change could then interfere with the request, like:
 - Request brokers and queue stage change to auth mode selection when done
 - Brokers received: automatic broker selection request is queued
 - Authentication starts
 - Stage change that was previously queued finally happens
 - We end up being back to auth mode selection

Also: https://github.com/ubuntu/authd/actions/runs/13813999841/job/38797674173

This has been, mostly, highlighted by the change to request the stage
change via messages, but we could have been wrong with the order of
other events anyways.

So, to prevent this to happen and to ensure that we follow the policy
that a stage doesn't do anything until it has the focus, delay the
automatic broker selection until the authentication selection stage is
actually focused.

See also commit 5133a87 that had similar rationale and also related to
the authentication selection procedure.