pam/authentication: Start the authentication only after the stage is focused
We used to start the authentication as soon as the model was filled,
but this could lead to start it before that the stage was actually
active (as being marked as focused).

This was not correct and might have lead to change stage events to
arrive later than authentication started events.

Make this order clearer now by only starting the authentication when the
stage has been focused and ensure this is always the case in tests.

In the CLI case now we've also to avoid writing the view if the
authentication has not been started yet, or the fields may not accept
input when they're about to start