testutils/daemon: Do not use the test parameter when sharing the daemon
If we're using a single shared daemon for all the tests the t* instance
may refer to another test that has been over, so don't use it once the
daemon has been started or we may end up in a race when logging.

In such case we use then other simpler logger and error checking functions.