pam/authentication: Don't use pointer-receivers for Update functions
The basic idea in the bubbletea model is that the Update function
changes the model and returns a copy of it, now we were not doing it in
the authentication model and in some authenticationComponent
implementations and this was leading to modifying some values in a racy
way, and indeed bubbletea was not ready for that as per the issue [1].

So, be consistent and don't use pointer receivers for Update functions
ever, so that we don't end up modifying the current model but rather a
copy of it. As it's expected to happen.

This required some extra love in newpasswordmodel because there we were
keeping track of the password entries pointers in a list and this list
now requires to be updated whenever one of the password entry that it
references has been udpated or its content won't never be exposed.

[1] https://github.com/charmbracelet/bubbletea/issues/909