pam/model: Do not use a pointer receiver for the UIModel Update
Like on the previous change we should not use pointer receiver for the
update functions, but rather act on a copy that we eventually return.

So that when the update is done the changes are applied.

This was not the case before and it was triggering some bubble tea races
so switch to use non-pointer receiver for the Update function.

This has the implication that the final exit status is never updated on
the model that gets initialized in the pam module, but we can now handle
this case by setting exit status on the program model just before
quitting.