pam/model: Implement the UIModel as proper bubbletea tea.Model
We used a pointer UIModel as the program model, but this was not how a
tea.Model was meant to be, as the whole struct was actually expected for
such purpose.

This was due to the fact that Init() had an input receiver, but again
this was not the way it should be, since Init is not meant to initialize
the content but rather to return messages to do that.

So cleanup the whole UIModel to split the creation with the init phases