pam/model: Make the uiModel private and only expose it as a tea.Model
Now that the model can be created with a specific function and that
it's properly implementing the tea.Model interface we can just expose it
externally as such, since there's no point to mess with its internal
structure from the module side.

This allows us to also not to export all the private variables.

A required change due to this was that now we need to rely again on
Update to stop the GDM conversations, but this is just cleaner as it
avoids again messing with the model in an unexpected way