ci/build-deb: Push the debian sources to packaging branches
In order to be able to build authd from launchpad recipes we need to use
a synchronized repository that contains the sources in a ready-to-build
form, however our sources requires vendording the dependencies and cargo
filtering and this does not work in builders without an internet
connection (and without very recent cargo either).

The good thing is that we're already building the sources ourselves, so
we can provide launchpad just them.

While dput'ing the built sources would be an option, it would also
requires having an uploader bot with its GPG key, and that's somewhat
unsafer, so let's instead just put the sources in a branch that is
updated at every push to main and that launchpad can fetch for creating
the packages.