ci/build-deb: Use the correct changes value to run the autopkgtests
github matrices are useful, but they don't provide a nice way to expose
their outputs so that subsequent jobs can reuse them easily.

In particular here the last job finishing was always replacing the
pkg-src-changes value and so that one was used in autopkg tests all the
times.

Thus, go a bit more manual here by defining the outputs per each job adding
the matrix-parameters and eventually read them as job env values.

This is not the best, but at least allows to separate the output values
by their ubuntu versions.

This does not affects the run-id value since it's the same for all the
jobs, so there's no need to set a specific one.