pam/integration-tests/ssh: Use NSS authd library in SSHd
We were running the tests simulating the existence of the users by using
a preloaded library that exposed fake users to sshd.

However, in authd we also have an NSS library that can do exactly the
same but exposing the users that authd and its broker expose.

So, let's do that:
 - Compile the NSS library and inject it to SSHd via preload
 - Update our wrapper so that always relies on the return value of the
   original getpwnam value, unless the requested user has to be always
   fake for some specific tests.
 - SSHD now fully relies on our library as in reality.