pam/integration-tests: Use pam_mkhomedir go create the user directories
As we do in our default pam module configuration, create the user $HOME
directory at session phase of our PAM stack.

This is handled by sshd through the very same pam_mkhomedir that is in
ubuntu and using its helper script.

However, we can't use the distro-provided module because that launches
the mkhomedir_helper binary without any environment set, and thus it
would ignore our preloaded NSS library, making the script to fail when
looking for the user directory.

So include in our code the very same implementation of the module that
is available upstream, with some tiny changes that are highlighted via
the AUTHD_TESTS_SSH_USE_AUTHD_NSS #ifdef.

The check that this works is in the tests, because once the ssh
connection is established the shell tries to cd into the $HOME and this
would fail if the module didn't work.