pam: Add utility function to safely log bubbletea messages
We need to log the tea messages but they may contain information that we
don't want to disclose.

So let's move the logging of the messages to a single place so that we
can avoid specific logic in each Update handler, and be consistent to
ensure that no message containing personal information won't be
disclosed.

At the same time such information is very useful when testing, so add
different implementations that are used when building in debug mode or
during tests.

Add also unit tests to ensure that the messages rewrite works