pam/integration-tests/gdm: Use gdm-way to change auth mode on the fly
Recently (mostly as per #827) we had an increase of gdm tests errors due
things such as:

    Diff:
    --- Expected
    +++ Actual
    @@ -1,9 +1 @@
    -{
    -  "type": "qrcode",
    -  "label": "Enter the following code after flashing the address: 1337",
    -  "button": "Regenerate code",
    -  "wait": "true",
    -  "entry": "",
    -  "content": "https://ubuntu.com",
    -  "code": ""
    -}
    +null
  Test: TestGdmModule/Authenticates_user_after_regenerating_the_qrcode_without_code_field
  Messages: Selected UI layout does not match expected one

The reason why it happened is that we were injecting an auth mode
selection just after a stage change request, without properly wait for
the change stage being completed, as the real GDM does.

So mimic the code that is used in GDM for real and by cancelling the
current authentication first and then selecting the new authentication
mode.

This is mimicking 1:1 what GDM does when an user requests an auth mode
selection change while a challenge is in progress, for example using the
gdm "key" button to select another mode without going back.

 See: https://github.com/ubuntu/authd/actions/runs/14315306989/job/40119981133