pam/integration-tests/gdm: Add more tests simulating auth mode selection switch
As per previous commits we're mostly only replicating what gdm does when
switching the authentication mode from the "key" button, and so from
challenge stage, without going back to the auth mode selection first.

Now, add tests for a more common scenario where a user may hit escape or
the "back" button to change the authentication mode.
And now do it safely, without requesting the auth mode selection too
early.

So, change the tests so that:
 - We get automatically redirected to password selection
 - Request a change stage to auth mode selection
 - Once back to auth mode selection, select the new auth mode
 - Use it to authenticate