userutils: Run the whole test inside bubblewrap
Move the helpers tools to run tests in bubblewrap to a more central
place (as they will surely be useful in future for other tests too).

Then run the whole userutils tests inside bubble wrap.

This is needed because we need to call lckpwdf(3) in the tests (see
the following commits), which writes to /etc (which we don't want to do
on the host, but in a bubblewrap environment).

And centralizing this allows us to just do such adjustments inside the
test that is launched in the bubblewrap environment, rather than on its
caller, simplifying the logic overall.

As per this we have also few other extra advantages:
 - Be able to easily add and run multiple tests inside the bwrap environment
 - Have a generic way to test inside a such environment (that will be useful
   in future for other integration tests)
 - Run all these tests in parallel