userutils: Run the whole test inside bubblewrap
Move the helpers tools to run tests in bubblewrap to a more central
place (as they will surely be useful in future for other tests too).

Then run the whole user util test inside bubble wrap.

In this way we don't have to simulate things outside, but just prepare
the environment for the test.

And it allows us to mock more things since the whole suite is running
inside a confined environment.