users/locking/tests: Add tests that try to lock an already locked database
If another process is locking the local accounts database authd will
hang until the calling process will not be done with it.

So ensure this is the case in tests. However, in order to verify this
we need to ensure that the locking process is not the same one already
having the lock, as in such case libc is nice enough to just return
early with an error.

So add a very simple locking test util program that will lock the db
indefinitely and then in our tests we can ensure that we can get back
the lock once that one is done.