Simplify propagation of context cancellation to D-Bus broker
D-Bus does not support cancelling ongoing method calls through context
cancellation, in contrast to gRPC. So if the context of the
IsAuthenticated call from the PAM module to authd is cancelled (e.g.
because the PAM module exited unexpectedly), we need propagate that to
the broker by calling the CancelIsAuthenticated method.

We used to do that by calling IsAuthenticated in a goroutine and
waiting for the context to be cancelled in the main goroutine, calling
CancelIsAuthenticated when the context is cancelled.

This commit simplifies that, by passing the context to BusObject.CallWithContext, and calling
when the returned error is context.Cancelled.

This is only done for the D-Bus broker, not the example broker. That's
not a problem, because the example broker already cancels the
IsAuthenticated call if the context is cancelled, without requiring a
separate CancelIsAuthenticated call.

This commit also returns auth.Cancelled if the context was cancelled,
which was actually never used with the actual D-Bus broker, but with the
example broker and the testutils.BrokerBusMock.